requirements.txt
google-api-core==2.12.0
google-api-python-client==2.101.0
google-auth==2.23.1
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
numpy==1.26.0
pandas==2.1.1
plotly==5.17.0
streamlit==1.27.0